ObjectSelectorControl with a bit of an easier way of setting options

+++ include/menuitems_object_multitoggle.h
@@ -131,7 +131,7 @@ class ObjectMultiToggleControl : public MenuItem {
                 char tmp[width_per_item];
                 for (uint8_t segment_start = 0 ; segment_start < strlen(item->label) ; segment_start += width_per_item-1) {
                     // copy width_per_item characters or max remaining
-                    int len_to_copy = min(strlen(item->label) - segment_start, width_per_item-1);
+                    int len_to_copy = min((int)(strlen(item->label) - segment_start), (int)(width_per_item-1));
 
                     //Serial.printf("\tcpos = %i, width_per_item = %i, len_to_copy = %i, strlen(item.label) = %i\n", segment_start, width_per_item, len_to_copy, strlen(item.label));
                     memcpy(&tmp[0], &item->label[segment_start], len_to_copy);
